Removing the file extension check on images
Since we're not checking mime types anyway, and most browser
supported image formats should probably render as attachments
on Mastodon and Pleroma, I'm removing this. There are often
normal JPEG and PNG images that get linked as an img src but
lack the standard file extensions, especially if it comes from
an API that returns the file binary.

Closes #36.